# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crow's Eye Marketing Platform API",
    "description": "REST API for the Crow's Eye marketing automation platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/stories/": {
      "post": {
        "tags": [
          "Stories"
        ],
        "summary": "Create Story",
        "description": "Create formatted story from media.\n\n**Tier Required:** Creator+ (Creator tier and above)\n\nAutomatically formats your media for Instagram/Facebook stories with:\n- Optimal aspect ratio (9:16)\n- Text overlays\n- Style templates\n- Transition effects",
        "operationId": "create_story_stories__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Stories"
        ],
        "summary": "List Stories",
        "description": "List user's stories.\n\n**Tier Required:** Creator+ (Creator tier and above)",
        "operationId": "list_stories_stories__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stories/{story_id}": {
      "get": {
        "tags": [
          "Stories"
        ],
        "summary": "Get Story",
        "description": "Get story by ID.\n\n**Tier Required:** Creator+ (Creator tier and above)",
        "operationId": "get_story_stories__story_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          },
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Stories"
        ],
        "summary": "Delete Story",
        "description": "Delete story by ID.\n\n**Tier Required:** Creator+ (Creator tier and above)",
        "operationId": "delete_story_stories__story_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          },
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StoryListResponse": {
        "properties": {
          "stories": {
            "items": {
              "$ref": "#/components/schemas/StoryResponse"
            },
            "type": "array",
            "title": "Stories"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "stories",
          "total"
        ],
        "title": "StoryListResponse",
        "description": "Story list response model."
      },
      "StoryRequest": {
        "properties": {
          "media_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Media Ids"
          },
          "text_overlay": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Overlay"
          },
          "style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style",
            "default": "modern"
          },
          "duration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration",
            "default": 15
          },
          "music": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Music",
            "default": false
          },
          "transitions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transitions",
            "default": "fade"
          }
        },
        "type": "object",
        "required": [
          "media_ids"
        ],
        "title": "StoryRequest",
        "description": "Story creation request model."
      },
      "StoryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "media_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Media Ids"
          },
          "text_overlay": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Overlay"
          },
          "style": {
            "type": "string",
            "title": "Style"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "preview_url": {
            "type": "string",
            "title": "Preview Url"
          },
          "download_url": {
            "type": "string",
            "title": "Download Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "media_ids",
          "text_overlay",
          "style",
          "duration",
          "created_at",
          "preview_url",
          "download_url"
        ],
        "title": "StoryResponse",
        "description": "Story response model."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}